mdl: import a single apply-to-all equation as ApplyToAll, not exploded slots
The importer expanded every subscripted MDL variable to the cartesian
product of its subscript elements, so a single apply-to-all equation
became N byte-identical Arrayed slots -- the one form in which a Vensim
dimension reference (DimA in an expression is the element's 1-based
position) has no active dimension to resolve against. Dimension
arithmetic in any arrayed MDL equation therefore failed to compile
through import (VECTOR ELM MAP(x[three], DimA - 1), spelled
byte-identically to the checked-in vector.mdl, while the XMILE twin ran
correctly), and every imported arrayed model carried N copies of each
A2A equation. A single covering apply-to-all equation now collapses to
Equation::ApplyToAll, the faithful translation -- the equivalence
harness had been compensating for the explosion by normalizing
identical slots to ApplyToAll for years.

The gate declines everything per-element for a reason that earned or
was pinned by a test: multiple source equations (including A2A over
different subranges), :EXCEPT:, an INITIAL equation or graphical
function (ApplyToAll has nowhere to carry either -- review caught the
collapse silently deleting an arrayed WITH LOOKUP's table), partial
cartesian coverage, and the whole opaque external-data placeholder
family (review caught GET XLS and kin regressing to EmptyEquation plus
CantSubscriptScalar at the diagnostic level while the file-level census
stayed clean; is_external_data_placeholder mirrors the writer's
is_data_equation enumeration). Census over the 262 MDL files under
test/: exactly one per-file change, vector.mdl compile-fail to ok, at
verdict and diagnostic-text level; the new simulates_vector_mdl_genuine
gates the fix against real-Vensim vector.dat. On C-LEARN the LTM var
count moves 6848 to 7163: measured strictly additive, all 315 the
GH #997 mapped per-element link scores this branch built, now reachable
because the aggregation variables arrive as ApplyToAll -- the emitted
pairings follow the declared 3-onto-7 element map (remaining_developed
is COP index 4, unreachable by any ordinal rule), and declines are
byte-identical. Census runs also surfaced a pre-existing
HashMap-order nondeterminism in which missing-file name an MDL import
error reports (GH #595's class), disclosed in the census harness. The
sketch-serialization block of writer_tests moves to a path-mounted
writer_sketch_tests sibling for the per-file line cap.
