+++ src/simlin-engine/src/mdl/CLAUDE.md
@@ -25,7 +25,7 @@ For design history and detailed implementation notes, see [docs/design/mdl-parse
 
 ### Conversion (`convert/`)
 - `mod.rs` -- Main conversion orchestration, group building, `DataProvider` threading
-- `variables.rs` -- Variable type detection (stock/flow/aux) and building; EXCEPT default_equation handling, GET DIRECT resolution
+- `variables.rs` -- Variable type detection (stock/flow/aux) and building; EXCEPT default_equation handling, GET DIRECT resolution. **A single apply-to-all MDL equation imports as one `Equation::ApplyToAll`**, not as N identical per-element slots: `build_variable_with_elements` expands every subscripted LHS to the cartesian product of its subscripts, so `y[DimA] = <rhs>` used to arrive as one slot per element all carrying the same `<rhs>`. That is the same equation written N times, and it is the form in which a DIMENSION reference cannot resolve -- Vensim's `DimA` in an expression is the element's 1-based position, and a per-element slot has no active apply-to-all dimension for it -- so `y[DimA] = VECTOR ELM MAP(x[three], (DimA - 1))`, legal Vensim and correct through the XMILE reader, failed to compile through this one. The collapse is gated by `slots_are_one_apply_to_all` plus a `single_apply_to_all` precondition: the source must be ONE equation with no `:EXCEPT:` (overrides and defaults are genuinely per-element); it must not be EXTERNAL-DATA-backed -- the whole opaque `{GET ...}` family, not just the `GET DIRECT` calls this module can resolve (`external_data::is_external_data_placeholder`), because a resolvable one's slots agree only as a property of the spreadsheet while an UNRESOLVABLE one leaves an empty equation in every slot, and collapsing that yields `ApplyToAll(dims, "")` -- an `EmptyEquation` error where the `Arrayed` form imported cleanly, taking its readers with it since a variable with no parseable equation has no dimensions for a consumer's `SUM(v[Dim!])`; no slot may carry an INITIAL equation or a graphical function (`ApplyToAll` is `(dims, equation)` and has nowhere to put either -- collapsing an arrayed `ACTIVE INITIAL` silently drops its initial, and collapsing an arrayed `WITH LOOKUP` or a bare arrayed lookup table DELETES the table outright, since it lives in the slots and not on the variable); and the slots must cover the dimensions' full cartesian product (a partial cover is not an apply-to-all). Not every clause is independently load-bearing, and the mutation matrix says which: dropping the INITIAL, graphical-function or coverage clause each reds its own tests, while the `:EXCEPT:`/default/equation-text clauses are belt-and-braces -- `needs_substitution` is exactly `expanded_eqs.len() > 1 || has_except_eq`, so when `single_apply_to_all` holds no per-element substitution runs and slot text agreement is guaranteed by construction. The rule is a STRICTER form of the one `src/simlin-engine/tests/integration/mdl_equivalence.rs`'s `normalize_equation` applies to compare us against xmutil (which emits apply-to-all here): that normalizer collapses on slot agreement alone, ignoring source-equation count, coverage and external-data backing, so anything the importer collapses the harness would too and the comparison stays consistent. Tests: `convert/apply_to_all_tests.rs` (collapse rows and must-not-collapse rows, each named for why its slots differ) and the corpus gate `simulate::simulates_vector_mdl_genuine`, which runs `vector.mdl` against real-Vensim `vector.dat` -- a twin that did not exist, which is why the defect survived as long as the fixture. Corpus-wide effect measured with `src/simlin-engine/examples/mdl_compile_census.rs`: across 262 MDL files, one model moved fail -> ok and none the other way.
 - `stocks.rs` -- Stock/flow linking via is_all_plus_minus algorithm
 - `dimensions.rs` -- Dimension/subscript building with range expansion and `DimensionMapping` construction
 - `external_data.rs` -- GET DIRECT DATA/CONSTANTS/LOOKUPS/SUBSCRIPT resolution via `DataProvider` trait
